Avoid dependency pollution from bookkeeper dependencies
If a module uses a class from a dependency, it should depend on that
dependency. That currently isn't the case with a lot of bookkeeper
dependencies, where code has been written to depend on classes which
are only pulled in transitively.

Removing the dependency pollution highlights a couple of issues. Junit
is being depended on, even though tests are TestNG. This dependency
should be removed. Also couple of places were using apache
commons-lang, even though pulsar doesn't declare this as a dependency
at all, but does declare commons-lang3.

This change updates the poms to directly depend anything the module
uses. The only code changes is to use common-lang3 instead of
common-lang, as the same classes are available in lang3.